Integrating a large number bug fixes, all of which (or most of which) are based on bug reports submitted through SourceForge:
Feature: The window's last-observed location is saved as a preference, and that location is used when the application starts later.

Bug fix: The Probe's Radix attribute displayed the values as they are saved in the file, rather than the display strings based on the current translation being used.

Appearance change: Rectangular XOR and XNOR gates are drawn with "=1" to conform to the ISO standard (previously it was simply "1"). However, if the "Multiple-Input Behavior" attribute is "When an odd number are on" and there are more than two inputs, the label is "2k+1" to reflect that it is actually doing an odd parity. (Previously it showed this label for odd parity even when there were just two inputs.)

Bug fix: Edits to the Bit Extender's Extension Type attribute or the Comparator's did not have an immediate effect in the circuit.

Bug fix: Under MacOS X, the title bar's close-window button would indicate that the file was changed after the first change, but the indicator would not disappear when the file was saved.

Bug fix: When a rectangle, oval, or rounded rectangle in a circuit's appearance had a paint type of "Border & Fill" and it was filled with solid black, that information was not properly stored in the saved file.

Bug fix: Libraries could not be unloaded.

Bug fix: The XNOR gate incorrectly did even parity by default, rather than being off only when exactly one input is one.

Bug fix: Most options were not loaded when a file was later reopened (all except the toolbar and mouse mappings).

Bug fix: The Logisim version saved in the project file was not interpreted correctly when opening the project. One consequence: If you edited the toolbar to include the Select and Wiring tools, they would be replaced by the Edit tool when reloaded.

Bug fix: When moving a selection so that one wire in the selection is moved exactly into the location of a wire being moved, the wire would sometimes disappear entirely rather than be moved. 

Bug fix: With the Edit Tool, pressing the backspace key deletes the current selection (just as it does when the delete key is pressed).

Bug fix: Edits to the Bit Extender's Extension Type attribute or the Comparator's did not have an immediate effect in the circuit.

Bug fix: Under MacOS X, the title bar's close-window button would indicate that the file was changed after the first change, but the indicator would not disappear when the file was saved.



git-svn-id: https://svn.code.sf.net/p/circuit/code/trunk@164 70edf91d-0b9e-4248-82e7-2488d7716404