Feature: Added a Wiring library containing transistor, transmission gate, power and ground components. Also, the splitter, pin, probe, tunnel, clock, pull resistor, and bit extender were moved into the Wiring library from the Base library, and constant was moved into the Wiring library from the Gates library.
Feature: In the Combinational Analysis module, the user can select between using a sum of products or a product of sums.

Bug fix: Changing the label for a Tunnel would not immediately affect the values propagated through the circuit.

Bug fix: In the Karnaugh maps drawn by the Combinational Analysis module, each implicant was indicated by a red rectangle, but they were meant to use varied colors.

Design change: Added a "Close" item to the Window menu, which closes the current window. The "Close" item in the File menu continues to close the current project; its shortcut is now Control-Shift-W.

Design change: For the Memory library's Random component, a reset when the seed is 0 computes a new initial seed based on the current time. (It previously reset to the initial seed for the current simulation.)


git-svn-id: https://svn.code.sf.net/p/circuit/code/trunk/src@181 70edf91d-0b9e-4248-82e7-2488d7716404