Bug fix: In saving customized circuit appearance in regions with a comma decimal separator, the comma would be saved into the file but not loaded properly. It now should always use a period for the decimal period - but it should also recover in cases where a comma is used instead.
Bug fix: After opening a dialog box for font or color and then changing the selection, the value seen from the dialog box continues to be shown in the attribute table, even if that is no longer the value associated with the selection.

Bug fix: In the Memory library, changing the Maximum Value of a Counter would not immediately be displayed in the attribute table.

git-svn-id: https://svn.code.sf.net/p/circuit/code/trunk@251 70edf91d-0b9e-4248-82e7-2488d7716404